A new filter name space has been created
task #53
