+++ tests/src/Orm/Events/Filter/AliasEntityCollectionTest.php
@@ -22,8 +22,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-namespace DSchoenbauer\Orm\Events\Validate\Schema;
+namespace DSchoenbauer\Orm\Events\Filter;
 
+use DSchoenbauer\Orm\Events\Filter\AliasEntityCollection;
 use DSchoenbauer\Orm\ModelInterface;
 use PHPUnit\Framework\TestCase;
