+++ tests/src/Orm/Events/Filter/DefaultValueTest.php
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-namespace DSchoenbauer\Orm\Events\Validate\Schema;
+namespace DSchoenbauer\Orm\Events\Filter;
 
 use DSchoenbauer\Orm\Entity\HasDefaultValuesInterface;
 use DSchoenbauer\Orm\Enum\ModelEvents;
